requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.23.5
pillow
matplotlib
transformers
opencv-python
nltk
streamlit